Compute pairwise relationship measures between all individuals in source and one individual in target

More explanation later.

comp_ind_pairwise(S, T, t, values, nGenos, Starts)

Arguments

S
"source", a matrix whose rows are integers, with NumInd-source rows and NumLoci columns, with each entry being a a base-0 representation of the genotype of the c-th locus at the r-th individual. These are the individuals you can think of as parents if there is directionality to the comparisons.

T
"target", a matrix whose rows are integers, with NumInd-target rows and NumLoci columns, with each entry being a a base-0 representation of the genotype of the c-th locus at the r-th individual. These are the individuals you can think of as offspring if there is directionality to the comparisons.

t
the index (base-1) of the individual in T that you want to compare against everyone on S.

values
the vector of genotype specific values. See the probs field of flatten_ckmr.

nGenos
a vector of the number of genotypes at each locus

Starts
the base0 indexes of the starting positions of each locus in probs.

Value

a data frame with columns "ind" (the base-1 index of the individual in S), "value" (the value extracted, typically a log likelihood ratio), and "num_loc" (the number of non-missing loci in the comparison.)