simulate a distribution of fake chromosome lengths

Sometimes you don't know the genome of your organism and you don't know where in the genome your markers sit. But you might have an idea of how big the genome is and how many chromosomes it is organized into. In that case, you can get a sense for how much an effect physical linkage will have on false positive rates, etc., by placing the markers randomly within a genome that approximates your organisms genome. The first step to that is simulating chromosome sizes. That is what this function does. It is very simple.

geometric_chromo_lengths(n, L, sl)

Arguments

n

number of chromosomes

L

length of genome in Gigabases

sl

length of the smallest chromosome as a fraction of the largest chromosome

Value

A list with three components:

chrom_lengths: a tibble with columns idx (chromsome index), chrom (chromosome name), scaled_length (length of the chromosome as a fraction of the whole genome), and num_bases. Note that the names of the chromosomes are like "fc01", "fc02", and so on. fc stands for "fake chromosome."

chrom_length_plot: a ggplot object that shows the simulated lengths of the chromosomes.

chrom_length_parameters: the parameters used in the function.

Examples

gcl <- geometric_chromo_lengths(
    n = 40,
    L = 4.6,
    sl = 0.5
)