compute the Y_l matrices for each locus in a list

Once you have the X_l matrices for each kappa, and the C_l matrices have been inserted into the list, then this function cycles over the loci and the kappa values and does the matrix multiplication to give you the matrix Y_l. This actually gives you the matrices Y_l_true and Y_l, but for now they are the same. I just want to put that into the code at this point so I can work with them appropriately.

insert_Y_l_matrices(L)

Arguments

L

the list of X_l and C_l matrices

Examples

example(insert_C_l_matrices)
#> 
#> in_C__> example(long_markers_to_X_l_list, package = "CKMRsim")
#> 
#> l___X_> data(kappas)
#> 
#> l___X_> lm_example <- long_markers_to_X_l_list(long_markers[1:18,], kappa_matrix = kappas)
#> 
#> l___X_> mh_example <- long_markers_to_X_l_list(microhaps, kappa_matrix = kappas)
#> 
#> in_C__> mh_cl_example <- insert_C_l_matrices(
#> in_C__+     mh_example,
#> in_C__+     ge_mod_assumed = ge_model_microhap1,
#> in_C__+     ge_mod_true = ge_model_microhap1
#> in_C__+ )
mh_yl_example <- insert_Y_l_matrices(mh_cl_example)