Microhaplotype data from kelp rockfish Sebastes atrovirens with pretend linkage positions

These are data from 165 amplicons sequenced in two different runs of an Illumina MiSeq machine at the SWFSC lab in Santa Cruz. Genotypes of roughly 150 individuals were called an resolved into haplotypes using the fact that each short read comes from a single chromosome. These were randomly assigned to some positions on some random chromosomes for testing linkage stuff.

linked_mhaps

Format

This tbl_df-ed data frame includes 825 alleles/haplotypes from 165 genomic regions and is in the format of long_markers. However since true positions of these markers in the genome are not known I just pretended that we knew some (randomly generated) locations. The column Chrom Is randomly given Locus has the name of each locus in it. The column Pos has a random position (in base pairs) for each locus. The Allele column gives the sequences of each allele/haplotype observed in the population.