Return locus-specific pairwise relationship measures between desired pairs of individuals

The idea here is that you can go back and look more closely at the log-likelihood ratios for pairs that are found to look the PO, etc., to see how much each of the different loci are contributing. More explanation later.

locus_specific_pairwise(S, T, s, t, values, nGenos, Starts)

Arguments

S

"source", a matrix whose rows are integers, with NumInd-source rows and NumLoci columns, with each entry being a a base-0 representation of the genotype of the c-th locus at the r-th individual. These are the individuals you can think of as parents if there is directionality to the comparisons.

T

"target", a matrix whose rows are integers, with NumInd-target rows and NumLoci columns, with each entry being a a base-0 representation of the genotype of the c-th locus at the r-th individual. These are the individuals you can think of as offspring if there is directionality to the comparisons.

s

a vector of base-1 indexes of the source individual in each pair.

t

a vector of base-1 indexes of the target individual in each pair. This vector is parallel to s. So, for example (s[i], t[i]) designates a pair that you wish to investigate (individual s[i] in S and t[i] in T)

values

the vector of genotype specific values. See the probs field of flatten_ckmr.

nGenos

a vector of the number of genotypes at each locus

Starts

the base0 indexes of the starting positions of each locus in probs.

Value

a data frame with columns "indS" (the base-1 index of the individual in S), "indT" (the base-1 index of the individual in S), "locus" (base-1 index of the locus), and "value" (the value extracted, typically a log likelihood ratio). If the pair is missing that locus it is given as NA_REAL