Model missing-data simulation results to interpolate FNR and FPR values

More later.

model_Qij_with_missing_data(
  Qtib,
  nu,
  de = "U",
  tr = "U",
  FNRs = c(0.3, 0.2, 0.1, 0.05, 0.01)
)

Arguments

Qtib

A tibble of missing-data simulation results, typically assembled from the Snakemake outputs created by summarize_missing_data(). It must include num_non_missing_loci, num_missing_loci, Qijs_unlinked, and Qijs_linked.

nu

the name of the relationship that is in the numerator of the likelihood ratio (Lambda) whose distribution you wish to learn about. It is a string, for example "FS", or "PO", or "U". The Q values for that relationship must be included in parameter Q. If this is a vector, then all different values are used in combination with all the values of de, tr, and, possibly, pstar. Corresponds to column "numerator" in the output

de

the relationship that appears in the denominator of Lambda. By default it is "U". Corresponds to column "denominator" in the output. If it is a vector, then all values are done iteratively in combination with other values as described for nu.

tr

the true relationship of the pairs. Default is "U". (i.e. you are going to get samples of Lambda under their distribution when the true relationship is tr). Operates over all values if a vector. Corresponds to column "true_relat" in the output.

FNRs

the false negative rates at which to evaluate the false positive rates. These are reported in column "fnr" in the output. These should all be between 0 and 1. By default fnr is c(0.3, 0.2, 0.1, 0.05, 0.01, 0.001).

Examples

# here for testing at the moment
#Qtib <- readr::read_rds("/tmp/Qtib.rds")