convert a long-format data frame of loci into a vector of lines for a Mendel def file

This makes a vector of text that has carriage returns and things in it as necessary. So, to include it in a def file, if the return from this function is called D you can just do cat(D, sep = "", file = myDef.txt), for example.

markers2mendel_def_lines(df)

Arguments

df

A data frame in the format of long_markers.

Details

Because it appears that Mendel requires unique locus names we will name them Chrom-underscore-Loc.

Examples

data(long_markers)
D <- markers2mendel_def_lines(long_markers)
cat(D[1:100], sep = "")
#> 4_1,,2 
#>    1,0.78994
#>    2,0.21006
#> 4_2,,4 
#>    1,0.48459
#>    2,0.23717
#>    3,0.16662
#>    4,0.11162
#> 4_3,,3 
#>    1,0.49899
#>    2,0.40213
#>    3,0.09888
#> 4_4,,4 
#>    1,0.40245
#>    2,0.32596
#>    3,0.15363
#>    4,0.11796
#> 4_5,,5 
#>    1,0.68270317296827
#>    2,0.145658543414566
#>    3,0.115348846511535
#>    4,0.03289967100329
#>    5,0.023389766102339
#> 4_6,,7 
#>    1,0.279262792627926
#>    2,0.223662236622366
#>    3,0.169411694116941
#>    4,0.117241172411724
#>    5,0.0877508775087751
#>    6,0.065950659506595
#>    7,0.056720567205672
#> 4_7,,6 
#>    1,0.455615443845562
#>    2,0.238227617723823
#>    3,0.15059849401506
#>    4,0.0875691243087569
#>    5,0.0496095039049609
#>    6,0.018379816201838
#> 4_8,,3 
#>    1,0.44874
#>    2,0.30166
#>    3,0.2496
#> 4_9,,8 
#>    1,0.28534
#>    2,0.2048
#>    3,0.20349
#>    4,0.1077
#>    5,0.08519
#>    6,0.06572
#>    7,0.03087
#>    8,0.01689
#> 4_10,,4 
#>    1,0.492635073649264
#>    2,0.2909970900291
#>    3,0.120488795112049
#>    4,0.0958790412095879
#> 4_11,,6 
#>    1,0.335043350433504
#>    2,0.286792867928679
#>    3,0.182271822718227
#>    4,0.0811308113081131
#>    5,0.0607606076060761
#>    6,0.0540005400054001
#> 4_12,,7 
#>    1,0.31895
#>    2,0.24648
#>    3,0.17771
#>    4,0.1275
#>    5,0.11007
#>    6,0.01726
#>    7,0.00203
#> 4_13,,4 
#>    1,0.507084929150708
#>    2,0.269787302126979
#>    3,0.136178638213618
#>    4,0.0869491305086949
#> 4_14,,6 
#>    1,0.30185
#>    2,0.2008
#>    3,0.19534
#>    4,0.1375
#>    5,0.08624
#>    6,0.07827
#> 4_15,,2 
#>    1,0.58465
#>    2,0.41535
#> 4_16,,3 
#>    1,0.45870458704587
#>    2,0.274322743227432
#>    3,0.266972669726697
#> 4_17,,4 
#>    1,0.681956819568196
#>    2,0.223692236922369
#>    3,0.0557405574055741
#>    4,0.038610386103861
#> 4_18,,2 
#>    1,0.54168
#>    2,0.45832
#> 4_19,,2 
#>    1,0.68136
#>    2,0.31864
#> 4_20,,2 
#>    1,0.51962
#>    2,0.48038
#> 4_21,,5 
#>    1,0.394683946839468
#>    2,0.246962469624696
#>    3,0.154371543715437
#>    4,0.118861188611886
#>    5,0.0851208512085121
#> 4_22,,6 
#>    1,0.343263432634326
#>    2,0.187531875318753
#>    3,0.160641606416064
#>    4,0.13220132201322
#>    5,0.107311073110731
#>    6,0.0690506905069051
#> 4_23,,7 
#>    1,0.279687203127969
#>    2,0.237167628323717
#>    3,0.176948230517695
#>    4,0.157178428215718
#>    5,0.124368756312437