create a vector of Mendel formatted map file strings

This makes a vector of text that has carriage returns and things in it as necessary. So, to include it in a map file, if the return from this function is called D you can just do cat(D, sep = "", file = myMap.txt), for example.

markers2mendel_map_lines(df)

Arguments

df

A data frame in the format of long_markers.

Examples

data(long_markers)
D <- markers2mendel_map_lines(long_markers)
cat(D[1:100], sep = "")
#> 4_1, 3845
#> 4_2, 71520
#> 4_3, 105104
#> 4_4, 256481
#> 4_5, 273360
#> 4_6, 390274
#> 4_7, 392419
#> 4_8, 404700
#> 4_9, 451714
#> 4_10, 635842
#> 4_11, 737459
#> 4_12, 806861
#> 4_13, 812460
#> 4_14, 861352
#> 4_15, 893394
#> 4_16, 1060464
#> 4_17, 1093048
#> 4_18, 1127711
#> 4_19, 1191102
#> 4_20, 1209795
#> 4_21, 1256130
#> 4_22, 1257931
#> 4_23, 1270541
#> 4_24, 1284443
#> 4_25, 1328475
#> 4_26, 1388581
#> 4_27, 1413632
#> 4_28, 1709922
#> 4_29, 1727510
#> 4_30, 1741347
#> 4_31, 1805626
#> 4_32, 1849959
#> 4_33, 1856567
#> 4_34, 1881739
#> 4_35, 1992783
#> 4_36, 2039308
#> 4_37, 2055524
#> 4_38, 2068442
#> 4_39, 2082924
#> 4_40, 2182021
#> 4_41, 2196153
#> 4_42, 2333651
#> 4_43, 2356155
#> 4_44, 2450593
#> 4_45, 2480089
#> 4_46, 2499318
#> 4_47, 2554832
#> 4_48, 2626596
#> 4_49, 2646648
#> 4_50, 2718367
#> 4_51, 2790664
#> 4_52, 2834038
#> 4_53, 2839716
#> 4_54, 2898480
#> 4_55, 2905930
#> 4_56, 2939583
#> 4_57, 2947693
#> 4_58, 2955607
#> 4_59, 2977592
#> 4_60, 2979963
#> 4_61, 3035392
#> 4_62, 3118270
#> 4_63, 3185330
#> 4_64, 3198657
#> 4_65, 3205089
#> 4_66, 3228022
#> 4_67, 3247352
#> 4_68, 3315355
#> 4_69, 3434058
#> 4_70, 3445110
#> 4_71, 3446368
#> 4_72, 3492570
#> 4_73, 3511348
#> 4_74, 3518544
#> 4_75, 3590884
#> 4_76, 3592878
#> 4_77, 3674768
#> 4_78, 3692505
#> 4_79, 3695292
#> 4_80, 3705134
#> 4_81, 3729078
#> 4_82, 3748918
#> 4_83, 3789745
#> 4_84, 3790282
#> 4_85, 3801425
#> 4_86, 3826161
#> 4_87, 3855889
#> 4_88, 3876502
#> 4_89, 3877627
#> 4_90, 3983031
#> 4_91, 4049150
#> 4_92, 4073769
#> 4_93, 4130606
#> 4_94, 4181557
#> 4_95, 4206704
#> 4_96, 4222611
#> 4_97, 4365019
#> 4_98, 4464504
#> 4_99, 4470814
#> 4_100, 4576879